Limelight Digital Bid Adapter: Backport latest changes to 4.43.x version branch
